requirements.txt
flask
gunicorn
pillow
python-dotenv
torch
torchvision
transformers
numpy
